requirements.txt
Django==2.0
django-filter==0.13.0
djangorestframework==3.7.3
mysqlclient==1.3.7
PyMySQL==0.8.0
